+++ src/ablDataDictionary.ts
@@ -9,10 +9,11 @@ export function openDataDictionary(project: OpenEdgeProjectConfig) {
     const env = process.env;
     env.DLC = project.dlc;
 
+    const currProfile = project.profiles.get(project.activeProfile);
     const prmFileName = path.join(tmpdir(), 'datadict-' + crypto.randomBytes(16).toString('hex') + '.json');
     const cfgFile = {
         verbose: false,
-        databases: project.dbConnections,
+        databases: currProfile.dbConnections,
         propath: [],
         parameters: [],
         returnValue: '',
@@ -23,6 +24,6 @@ export function openDataDictionary(project: OpenEdgeProjectConfig) {
     fs.writeFileSync(prmFileName, JSON.stringify(cfgFile));
     const prms = ["-clientlog", path.join(project.rootDir, ".builder\\dictionary.log"), "-p", path.join(__dirname, '../resources/abl-src/dynrun.p'), "-param", prmFileName, "-basekey", "INI", "-ininame", path.join(__dirname, '../resources/abl-src/empty.ini')];
 
-    cp.spawn(project.getExecutable(true), project.extraParameters.split(' ').concat(prms), { env: env, cwd: project.rootDir, detached: true });
+    cp.spawn(currProfile.getExecutable(true), currProfile.extraParameters.split(' ').concat(prms), { env: env, cwd: project.rootDir, detached: true });
 }